requirements.txt
# requirements.txt
alpaca-py==0.10.0
python-dotenv==1.0.0
pandas==2.1.0